requirements.txt
yt-dlp
faster-whisper
sentence-transformers
faiss-cpu
ollama
streamlit
gtts
numpy